pyproject.toml
[project]
name = "toweroffantasy.api"
version = "0.1.0"
description = "Default template for PDM package"
authors = [
    { name = "Gabriel Nobrega", email = "gabrieltkdnobrega63@gmail.com" },
]
dependencies = [
    "fastapi[standard]>=0.115.3",
    "pydantic>=2.9.2",
    "pydantic-settings>=2.6.0",
    "pyjwt>=2.9.0",
    "bcrypt>=4.2.0",
    "psycopg[binary]>=3.2.3",
    "redis[hiredis]>=5.2.0",
    "aiofiles>=24.1.0",
    "unidecode>=1.3.8",
    "prisma>=0.15.0",
]
requires-python = "==3.12.*"
readme = "README.md"
license = { file = "LICENSE" }


[tool.pdm]
distribution = false

[tool.pdm.scripts]
start = { cmd = "uvicorn src.main:create_app --factory --host 0.0.0.0 --port 8000", env_file = ".env" }
dev = { composite = ["setup-docker", "setup-dev"] }

setup-docker = { cmd = "docker compose up -d" }
setup-dev = { cmd = "uvicorn src.main:create_app --factory --reload --reload-dir src", env_file = ".env" }

migrate = { cmd = "prisma migrate dev", env_file = ".env" }
generate = { cmd = "prisma generate", env_file = ".env" }

[dependency-groups]
dev = ["ruff>=0.8.4", "types-pyjwt>=1.7.1", "pre-commit>=4.0.1"]
